Also add the CLEAN command.

+++ BUILD
@@ -12,16 +12,26 @@ if [[ "${@: -1}" == "--verbose" ]]; then
   set -- "${@:1:$(($#-1))}"  # Remove the last argument (--verbose)
 fi
 
-
 # Initialize and update git submodules
 if [ -d .git ]; then
   git submodule init
   git submodule update
 fi
 
-
 # Define the directory for local Miniconda installation
 MINICONDA_DIR=".deploy/Miniconda3"
+
+# Check if .deploy/conda_environment/ exists
+if [ -d ".deploy/conda_environment/" ]; then
+  echo ".deploy/conda_environment/ exists. CLEAN first. Exiting."
+  exit 1
+fi
+
+# Check if $MINICONDA_DIR exists
+if [ -d "$MINICONDA_DIR" ]; then
+  echo "$MINICONDA_DIR exists. CLEAN first. Exiting."
+  exit 1
+fi
 
 # Clean up any previous Conda environment and build targets
 rm -rf .deploy/conda_environment/
@@ -34,15 +44,6 @@ rm /tmp/Miniconda3.sh
 
 # Initialize conda in the current shell (without modifying any shell configuration files)
 eval "$(.deploy/Miniconda3/bin/conda shell.bash hook)"
-
-
-
-
-# Activate the base Conda environment
-conda activate base
-
-# Clean up any previous Conda environment and build targets
-rm -rf .deploy/conda_environment/
 
 # Create and activate a new Conda environment based on .deploy/conda.yaml configuration
 conda env create --prefix .deploy/conda_environment/ --file .deploy/conda.yaml
